+++ packages/pwa/src/components/nav/LeftSidebar.css.ts
