+++ packages/pwa/src/screens/404.tsx
@@ -1,23 +1,20 @@
-import {NavItems} from '@shared/lib/navItems.shared';
+import {DEFAULT_NAV_ITEM} from '@shared/lib/navItems.shared';
 
-import {NavItemId} from '@shared/types/urls.types';
-
-import {Link} from '@src/components/atoms/Link';
 import {Text} from '@src/components/atoms/Text';
-
-import {rootRoute} from '@src/routes/__root';
+import {NavItemLink} from '@src/components/nav/NavItemLink';
 
 export const NotFoundScreen: React.FC<{readonly message: string}> = ({message}) => {
-  const navItem = NavItems.fromId(NavItemId.Untriaged);
+  const navItem = DEFAULT_NAV_ITEM;
+
   return (
     <div className="flex h-full w-full flex-col items-center justify-center gap-2">
       <Text as="h1">404</Text>
       <Text as="p">{message}</Text>
-      <Link to={rootRoute.fullPath}>
+      <NavItemLink navItemId={navItem.id}>
         <Text as="p" underline="always">
           Go to {navItem.title}
         </Text>
-      </Link>
+      </NavItemLink>
     </div>
   );
 };
