+++ packages/pwa/src/screens/ErrorScreen.tsx
@@ -1,25 +1,22 @@
-import {NavItems} from '@shared/lib/navItems.shared';
+import {DEFAULT_NAV_ITEM} from '@shared/lib/navItems.shared';
 
-import {NavItemId} from '@shared/types/urls.types';
-
-import {Link} from '@src/components/atoms/Link';
 import {Text} from '@src/components/atoms/Text';
-
-import {rootRoute} from '@src/routes/__root';
+import {NavItemLink} from '@src/components/nav/NavItemLink';
 
 // TODO: Improve design of error screen.
 export const ErrorScreen: React.FC<{readonly error: Error}> = ({error}) => {
-  const navItem = NavItems.fromId(NavItemId.Untriaged);
+  const navItem = DEFAULT_NAV_ITEM;
+
   return (
     <div className="flex h-full w-full items-center justify-center">
       <div className="flex max-w-[960px] flex-col items-center gap-8">
         <Text as="h1">Oops... something went wrong</Text>
         <Text as="p">{error.message}</Text>
-        <Link to={rootRoute.fullPath}>
+        <NavItemLink navItemId={navItem.id}>
           <Text as="p" underline="always">
             Go to {navItem.title}
           </Text>
-        </Link>
+        </NavItemLink>
       </div>
     </div>
   );
